==> 1.1_diffs.ipynb <==
import pandas as pd

vcf_cols = ['CHROM','POS','ID','REF','ALT','QUAL','FILTER','INFO','FORMAT','disk/data/dic_svevo.region.bam']

#only_lan_svevo
file_only_in = 'disk/data/variant_lan_svevo.ann.vcf'
file_compare_to = 'disk/data/variant_dic_svevo.ann.vcf'
file_res = 'disk/data/variants_only_lan_svevo.ann.vcf.csv'

#only_dic_svevo
file_only_in = 'disk/data/variant_dic_svevo.ann.vcf'
file_compare_to = 'disk/data/variant_lan_svevo.ann.vcf'
file_res = 'disk/data/variants_only_dic_svevo.ann.vcf.csv'

#only_dic_wew
file_only_in = 'disk/data/variant_dic_wew.ann.vcf'
file_compare_to = 'disk/data/variant_lan_wew.ann.vcf'
file_res = 'disk/data/variants_only_dic_wew.ann.vcf.csv'

#only_lan_wew
file_only_in = 'disk/data/variant_lan_wew.ann.vcf'
file_compare_to = 'disk/data/variant_dic_wew.ann.vcf'
file_res = 'disk/data/variants_only_lan_wew.ann.vcf.csv'

df_in = pd.read_csv(file_only_in, sep='\t', comment='#', header=None)
df_in.columns = vcf_cols
print(len(df_in.index))

df_to = pd.read_csv(file_compare_to, sep='\t', comment='#', header=None)
df_to.columns = vcf_cols
print(len(df_to.index))

count_both = count_diff = 0
r_variants = []
for k,v in df_in.iterrows(): #(R)
    res = df_to[(df_to.CHROM == v.CHROM) & (df_to.POS == v.POS) & (df_to.REF == v.REF)  & (df_to.ALT == v.ALT)]
    if(len(res.index) > 0): #the (S) genotype has the same variant?
        count_both += 1
    else:
        count_diff += 1
        r_variants.append(v) #this variant is only present in (R)
print('Variants in both genotypes: ' + str(count_both))
print('Unique variants: ' + str(count_diff))


df = pd.DataFrame(r_variants)
df.head(3)

df.to_csv(file_res, sep='\t', index=False)

df[df.INFO.str.contains('HIGH')]

df[df.INFO.str.contains('MODIFIER')]
