==> 2.2_blast_crossref.ipynb <==
"""
Blasts genes from svevo and zavitan and finds putative orthologues
Outputs first, a table with genes corresponding to each reference
Adds interesting information to output, position in genome according to annotation,
position of blast against genome, GO:terms
"""
#blastn -perc_identity 85 -subject disk/counts/svevo.counts.fa -query disk/counts/zav_v2.counts.fa -outfmt "6 qseqid sseqid pident mismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/blast_zv_sv.csv
#blastn -perc_identity 85 -query disk/counts/svevo.counts.fa -subject disk/counts/zav_v2.counts.fa -outfmt "6 qseqid sseqid pident mismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/blast_sv_zv.csv



import pandas as pd

file_zavitan = 'disk/counts/zavitan_v2.genes.csv'
file_svevo = 'disk/counts/svevo.genes.csv'
file_blast_sv_zv = 'disk/counts/blast_sv_zv.csv' #we'll use this layout
file_blast_zv_sv = 'disk/counts/blast_zv_sv.csv' #we'll use this layout

df_zavitan_all = pd.read_csv(file_zavitan, sep=',', )
df_svevo_all = pd.read_csv(file_svevo, sep=',', )

df_blast_sv_zv = pd.read_csv(file_blast_sv_zv, sep='\t', header=None)
df_blast_sv_zv.columns = ['svevo','zavitan','pident','mismatch','gapopen','qstart','qend','sstart','send','evalue','length','qlen','slen']


df_blast_zv_sv = pd.read_csv(file_blast_zv_sv, sep='\t', header=None)
df_blast_zv_sv.columns = ['zavitan','svevo','pident','mismatch','gapopen','qstart','qend','sstart','send','evalue','length','qlen','slen']


df_blast = df_blast[((df_blast.length / df_blast.qlen) >= 0.7) &
       ((df_blast.length / df_blast.slen) >= 0.7) &
       ((df_blast.length / df_blast.qlen) <= 1.3) &
       ((df_blast.length / df_blast.slen) <= 1.3)]

print('blast:',len(df_blast.index))
df_blast.tail(4)

df_zavitan = df_zavitan_all[['id']]
df_svevo = df_svevo_all[['id']]

df_zavitan_merge = pd.merge(df_zavitan, df_blast, left_on='id', right_on='zavitan', how='left')
df_zavitan_merge = df_zavitan_merge[['id','svevo']]
df_zavitan_merge.head(3)

df_svevo_merge = pd.merge(df_svevo, df_blast, left_on='id', right_on='svevo', how='left')
df_svevo_merge = df_svevo_merge[['id','zavitan']]
df_svevo_merge.head(3)

df_merge = pd.merge(df_svevo_merge, df_zavitan_merge, left_on='id', right_on='svevo', how='outer')
df_merge = df_merge[['id_x','id_y']]
df_merge.columns = ['svevo','zavitan']


df_merge.to_csv('disk/counts/cross.csv', sep=',', index=None)

print(len(df_merge.index))
df_merge.head(3)

#add info to list
file_ann_svevo = 'disk/PGSB_TRITD_Jan2017_all.gff3'
file_ann_zav_v2 = 'disk/SORTED_WEW_v2_HC_e_LC_GFF3_CATTATI_gff_PAO_26feb18.gff3'

df_ann_sv = pd.read_csv(file_ann_svevo, index_col=False, sep='\t', header=None)
df_ann_sv.columns = ['seqname', 'source', 'feature', 'start', 'end', 'score', 'strand', 'frame', 'attribute']

df_ann_zv = pd.read_csv(file_ann_zav_v2, index_col=False, sep='\t', header=None)
df_ann_zv.columns = ['seqname', 'source', 'feature', 'start', 'end', 'score', 'strand', 'frame', 'attribute']

df_ann_zv = df_ann_zv[df_ann_zv.feature=='gene']
df_ann_zv = df_ann_zv[df_ann_zv.seqname == 'chr3A']
df_ann_sv = df_ann_sv[df_ann_sv.feature =='gene']
df_ann_sv = df_ann_sv[df_ann_sv.seqname == 'chr3A']



df_ann_sv['gene'] = df_ann_sv.attribute.str.split(';', expand=True)[0].str.split('=', expand=True)[1]
df_ann_zv['gene'] = df_ann_zv.attribute.str.split(';', expand=True)[0].str.split('=', expand=True)[1]

df_ann_sv = df_ann_sv[['seqname','start','end','gene']]
df_ann_zv = df_ann_zv[['seqname','start','end','gene']]

#adds svevo coordinates
df_merge_svevo = pd.merge(df_merge, df_ann_sv, left_on='svevo', right_on='gene', how='left')
df_merge_svevo = df_merge_svevo[['seqname','start','end','svevo','zavitan']]
df_merge_svevo.columns = ['svevo.seqname','svevo.start','svevo.end','svevo.gene','zavitan.gene']

#adds zavitan coordinates
df_merge_zavitan = pd.merge(df_merge_svevo, df_ann_zv, left_on='zavitan.gene', right_on='gene', how='left')
df_merge_zavitan = df_merge_zavitan[['svevo.seqname','svevo.start','svevo.end','svevo.gene','zavitan.gene','seqname','start','end']]
df_merge_zavitan.columns = ['svevo.seqname','svevo.start','svevo.end','svevo.gene','zavitan.gene','zavitan.seqname','zavitan.start','zavitan.end']



print(len(df_merge_zavitan.index))
df_merge_zavitan.drop_duplicates( keep='first', inplace=True)
print("W/O duplicates:",len(df_merge_zavitan.index))

df_merge_zavitan.head(4)




df_merge_zavitan = pd.merge(df_merge_zavitan,df_svevo_all[['id','log2FoldChange','padj']], left_on='svevo.gene', right_on='id', how='left')
df_merge_zavitan.rename(columns= {'log2FoldChange':'svevo.l2fc', 'padj':'svevo.padj'}, inplace=True)
df_merge_zavitan.drop('id', axis=1, inplace=True)


df_merge_zavitan = pd.merge(df_merge_zavitan,df_zavitan_all[['id','log2FoldChange','padj']], left_on='zavitan.gene', right_on='id', how='left')
df_merge_zavitan.rename(columns={'log2FoldChange':'zavitan.l2fc', 'padj':'zavitan.padj'}, inplace=True)
df_merge_zavitan.drop('id', axis=1, inplace=True)


df_merge_zavitan = df_merge_zavitan[['svevo.seqname','svevo.start','svevo.end','svevo.padj','svevo.l2fc','svevo.gene','zavitan.gene','zavitan.seqname','zavitan.start','zavitan.end','zavitan.padj','zavitan.l2fc']]
df_merge_zavitan.to_csv('disk/counts/cross_info.csv', sep=',', index=None)
