==> 4.1_blast genes in other genome.ipynb <==
blastn  -perc_identity 85 -query disk/counts/zav_v2.counts.fa    -subject disk/counts/svevo3A.fasta  -outfmt "6 qseqid sseqid pidentmismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/zavitan_genes_svevo_3A.csv
blastn  -perc_identity 85 -query disk/counts/zav_v2.counts.fa    -subject disk/counts/svevo.un.fa  -outfmt "6 qseqid sseqid pidentmismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/zavitan_genes_svevo_UN.csv
blastn  -perc_identity 85 -query disk/counts/svevo.counts.fa    -subject disk/counts/zavitan.un.fa  -outfmt "6 qseqid sseqid pidentmismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/svevo_genes_zavitan_UN.csv
blastn  -perc_identity 85 -query disk/counts/svevo.counts.fa    -subject disk/counts/zavitan_3A.fasta  -outfmt "6 qseqid sseqid pidentmismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/svevo_genes_zavitan_3A.csv
blastn  -perc_identity 85 -query disk/counts/svevo.counts.fa    -subject disk/counts/zavitan_3A.fasta  -outfmt "6 qseqid sseqid pident mismatch gapopen qstart qend sstart send evalue length qlen slen"  -evalue 10e-10 > disk/counts/svevo_genes_zavitan_3A.csv


import pandas as pd

#ZV genes in SV genome
file_blast = 'disk/counts/disk/counts/zavitan_genes_svevo_genome.csv'


#SV genes in ZV genome
file_blast = 'disk/counts/svevo_genes_zavitan_genome.csv'
file_cross = 'disk/counts/cross_info.csv'

df_blast = pd.read_csv(file_blast, sep=',', header=None)
df_blast.columns = ['qseqid','sseqid','pident','mismatch','gapopen','qstart','qend','sstart','send','evalue','length','qlen','slen']
df_blast.head(4)

df_blast = df_blast[((df_blast.length / df_blast.qlen) >= 0.7) &
       ((df_blast.length / abs(df_blast.sstart - df_blast.send)) >= 0.7) &
       ((df_blast.length / df_blast.qlen) <= 1.3) &
       ((df_blast.length / abs(df_blast.sstart - df_blast.send)) <= 1.3)]
print(len(df_blast.index))

df_blast.head(3)

#
df_cross = pd.read_csv(file_cross, sep='\t', )
df_cross.head(2)
